Speed up deployments by avoiding re-rsync of webtree if DRUPAL_REBUILD_ON_REDEPLOY not TRUE

+++ init/60_build_drupal_tree.sh
@@ -23,17 +23,22 @@ then
 elif [[ -f /tmp/DRUPAL_DB_LIVE && -f /tmp/DRUPAL_FILES_LIVE ]];
 then
   # Site Needs Upgrade
-  echo "Database Exists and Files Found, Updating Existing Site"
+  echo "Database Exists and Files Found. DRUPAL_REBUILD_ON_REDEPLOY=$DRUPAL_REBUILD_ON_REDEPLOY"
+
+  if [[ "$DRUPAL_REBUILD_ON_REDEPLOY" == "TRUE" ]];
+  then
+    echo "Updating Existing Site.."
+
+    # Transfer the pre-built drupal tree.
+    rsync -a --stats --progress --no-perms --no-owner --no-group --delete --exclude=sites/default/files/ --exclude=sites/default/settings.php ${DRUSH_MAKE_TMPROOT}/ ${DRUPAL_ROOT}/
+
+    # Apply database updates, if they exist.
+    drush --yes --root=${DRUPAL_ROOT} --uri=default updb
+  fi
 
   # Ensure the database details are still valid.
   sed -i "s|'host' => '[0-9]\{1,3\}\.[0-9]\{1,3\}\.[0-9]\{1,3\}\.[0-9]\{1,3\}',|'host' => '$MYSQL_PORT_3306_TCP_ADDR',|g" ${DRUPAL_ROOT}/sites/default/settings.php
   sed -i "s|'port' => '[0-9]\{2,4\}',|'port' => '$MYSQL_PORT_3306_TCP_PORT',|g" ${DRUPAL_ROOT}/sites/default/settings.php
-
-  # Transfer the pre-built drupal tree.
-  rsync -a --stats --progress --no-perms --no-owner --no-group --delete --exclude=sites/default/files/ --exclude=sites/default/settings.php ${DRUSH_MAKE_TMPROOT}/ ${DRUPAL_ROOT}/
-
-  # Apply database updates, if they exist.
-  drush --yes --root=${DRUPAL_ROOT} --uri=default updb
 
 else
   # Inconsistency detected, do nothing to avoid data loss.
